Named params api changed
Now the IntoParams#to_params return a ParamsType, instead of
Vec<IntoParam>. The ParamsType has two options: named and unamed. All
default implementations of IntoParams use the unamed option and the
derive use named.

With the named option, the derive macro implementation provide a hashmap
with all fields names and his values.

For cover all the lib, the NamedParams api are called directly in the
statement module