Event working
Event API working with native client.

The events mannager now use the connection builder as a param
to spawn a new connection for every event added instead of using
a single one.

To control the events, we use some threads. After the start() call
the user needs call the join() thread method.

Issue #133